Split up the bind.py file
